% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_oeis_ids}
\alias{extract_oeis_ids}
\title{Extract OEIS IDs from text}
\usage{
extract_oeis_ids(text)
}
\arguments{
\item{text}{Character string to search}
}
\value{
Character vector of OEIS IDs
}
\description{
Extract OEIS IDs from text
}
\examples{
extract_oeis_ids("Cf. A000045, A000032 and A000045.")

}
