% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfile.R
\name{get_bfile_data}
\alias{get_bfile_data}
\title{Get B-file Data}
\usage{
get_bfile_data(bfile)
}
\arguments{
\item{bfile}{A BFile object}
}
\value{
A \code{gmp::bigz} vector or NULL
}
\description{
Get B-file Data
}
\examples{
\donttest{
primes <- BFile("A000040")
head(get_bfile_data(primes))
}

}
