% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_bfile_info}
\alias{get_bfile_info}
\title{Get B-file Information}
\usage{
get_bfile_info(seq)
}
\arguments{
\item{seq}{A Sequence object}
}
\value{
List with b-file metadata: available, filename, url, length,
first, last, min, max
}
\description{
Get B-file Information
}
\examples{
\dontshow{if (oeis_available()) withAutoprint(\{ # examplesIf}
fib <- Sequence("A000045")
get_bfile_info(fib)
\dontshow{\}) # examplesIf}
}
