% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_bibtex}
\alias{get_bibtex}
\title{Build a BibTeX @misc entry citing this OEIS sequence}
\usage{
get_bibtex(seq)
}
\arguments{
\item{seq}{A Sequence object}
}
\value{
Character string with a BibTeX \verb{@misc} entry
}
\description{
Build a BibTeX @misc entry citing this OEIS sequence
}
\examples{
\dontshow{if (oeis_available()) withAutoprint(\{ # examplesIf}
fib <- Sequence("A000045")
cat(get_bibtex(fib))
\dontshow{\}) # examplesIf}
}
