% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfile.R
\name{get_filename}
\alias{get_filename}
\title{Get B-file Filename}
\usage{
get_filename(bfile)
}
\arguments{
\item{bfile}{A BFile object}
}
\value{
Character string
}
\description{
Get B-file Filename
}
\examples{
\dontshow{if (oeis_available()) withAutoprint(\{ # examplesIf}
primes <- BFile("A000040")
get_filename(primes)
\dontshow{\}) # examplesIf}
}
