% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oeis_bfile}
\alias{oeis_bfile}
\title{Generate OEIS b-file filename}
\usage{
oeis_bfile(oeis_id)
}
\arguments{
\item{oeis_id}{Character string (e.g., "A000045")}
}
\value{
Character string (e.g., "b000045.txt")
}
\description{
Generate OEIS b-file filename
}
\examples{
oeis_bfile("A000045")

}
