% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oeis_url}
\alias{oeis_url}
\title{Generate OEIS URL}
\usage{
oeis_url(oeis_id, fmt = NULL)
}
\arguments{
\item{oeis_id}{Character string (e.g., "A000045")}

\item{fmt}{Character string: "json", "text", "bfile", "graph", or NULL}
}
\value{
Character string containing the URL
}
\description{
Generate OEIS URL
}
\examples{
oeis_url("A000045")
oeis_url("A000045", fmt = "json")

}
