% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfile.R
\name{BFile}
\alias{BFile}
\title{OEIS B-file Class}
\usage{
BFile(oeis_id)
}
\arguments{
\item{oeis_id}{Character string (e.g., "A000045")}
}
\value{
An object of class "BFile"
}
\description{
OEIS B-file Class
}
