% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfile.R
\name{get_bfile_indices}
\alias{get_bfile_indices}
\title{Get B-file Indices}
\usage{
get_bfile_indices(bfile)
}
\arguments{
\item{bfile}{A BFile object}
}
\value{
Integer vector of the first-column indices, or NULL when parsing failed
}
\description{
Get B-file Indices
}
\examples{
\dontshow{if (oeis_available()) withAutoprint(\{ # examplesIf}
primes <- BFile("A000040")
head(get_bfile_indices(primes))
\dontshow{\}) # examplesIf}
}
