% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_data_values}
\alias{get_data_values}
\title{Return the parsed sequence data terms}
\usage{
get_data_values(seq)
}
\arguments{
\item{seq}{A Sequence object}
}
\value{
A \code{gmp::bigz} vector of terms
}
\description{
Return the parsed sequence data terms
}
\examples{
\donttest{
fib <- Sequence("A000045")
get_data_values(fib)[1:10]
}

}
