% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_keyword_description}
\alias{get_keyword_description}
\title{Return the OEIS keyword description for a given tag}
\usage{
get_keyword_description(seq, keyword_tag)
}
\arguments{
\item{seq}{A Sequence object}

\item{keyword_tag}{Character string}
}
\value{
Character string with description, or NULL
}
\description{
Instance-method wrapper around \code{\link[=oeis_keyword_description]{oeis_keyword_description()}}.
}
