% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfile.R
\name{get_url}
\alias{get_url}
\title{Get B-file URL}
\usage{
get_url(bfile)
}
\arguments{
\item{bfile}{A BFile object}
}
\value{
Character string
}
\description{
Get B-file URL
}
\examples{
\donttest{
primes <- BFile("A000040")
get_url(primes)
}

}
