% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_xref_ids}
\alias{get_xref_ids}
\title{Extract cross-reference IDs}
\usage{
get_xref_ids(seq)
}
\arguments{
\item{seq}{A Sequence object}
}
\value{
Character vector of OEIS IDs
}
\description{
Extract cross-reference IDs
}
