% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oeis_keyword_description}
\alias{oeis_keyword_description}
\title{Get OEIS keyword description}
\usage{
oeis_keyword_description(keyword_tag)
}
\arguments{
\item{keyword_tag}{Character string}
}
\value{
Character string with description or NULL
}
\description{
Get OEIS keyword description
}
