% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{OEIS_URL}
\alias{OEIS_URL}
\alias{oeis_url}
\title{OEIS URL constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
OEIS_URL

oeis_url(oeis_id, fmt = NULL)
}
\arguments{
\item{oeis_id}{Character string (e.g., "A000045")}

\item{fmt}{Character string: "json", "text", "bfile", "graph", or NULL}
}
\value{
Character string containing the URL
}
\description{
OEIS URL constant

Generate OEIS URL
}
\examples{
oeis_url("A000045")
oeis_url("A000045", fmt = "json")

}
\keyword{internal}
