% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{plot.Sequence}
\alias{plot.Sequence}
\title{Plot Sequence Data}
\usage{
\method{plot}{Sequence}(x, ...)
}
\arguments{
\item{x}{A Sequence object}

\item{...}{Additional arguments passed to plot_data}
}
\description{
Plot Sequence Data
}
