% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfile.R
\name{plot_data}
\alias{plot_data}
\title{Plot B-file Data}
\usage{
plot_data(bfile, n = NULL, plot_style = "line", ...)
}
\arguments{
\item{bfile}{A BFile object}

\item{n}{Number of points to plot}

\item{plot_style}{"line" or "scatter"}

\item{...}{Additional arguments}
}
\value{
A ggplot2 plot
}
\description{
Plot B-file Data
}
