% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{Sequence}
\alias{Sequence}
\title{OEIS Sequence Class}
\usage{
Sequence(oeis_id)
}
\arguments{
\item{oeis_id}{Character string (e.g., "A000045")}
}
\value{
An object of class "Sequence"

Fetches an OEIS entry (and its b-file) over the internet.
}
\description{
OEIS Sequence Class
}
\examples{
\donttest{
fib <- Sequence("A000045")
fib$name
}

}
