% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{Sequence}
\alias{Sequence}
\title{OEIS Sequence Class}
\usage{
Sequence(oeis_id)
}
\arguments{
\item{oeis_id}{Character string (e.g., "A000045")}
}
\value{
An object of class "Sequence"
}
\description{
OEIS Sequence Class
}
\examples{
\donttest{
seq <- Sequence("A000045")
print(seq$name)
}

}
